Code cleanup and refactor.
