+++ src/components/Assets/index.tsx
@@ -1,0 +1,4 @@
+import AssetDrawer from './AssetDrawer'
+import AssetsHeader from './AssetsHeader'
+
+export { AssetDrawer, AssetsHeader }
